move woff_to_otf function into opentype.js
