+++ src/opentype.js
@@ -9,6 +9,7 @@ import Font from './font.js';
 import Glyph from './glyph.js';
 import { CmapEncoding, GlyphNames, addGlyphNames } from './encoding.js';
 import parse from './parse.js';
+import { encode } from './types.js';
 import BoundingBox from './bbox.js';
 import Path from './path.js';
 import cpal from './tables/cpal.js';
@@ -35,7 +36,6 @@ import os2 from './tables/os2.js';
 import post from './tables/post.js';
 import meta from './tables/meta.js';
 import gasp from './tables/gasp.js';
-import { woff_to_otf } from './woff-to-otf.js';
 /**
  * The opentype library.
  * @namespace opentype
@@ -534,6 +534,87 @@ function loadSync(url, opt) {
     return parseBuffer(require('fs').readFileSync(url), opt);
 }
 
+/**
+ * Convert/Uncompress a buffer of a woff font to otf/ttf without parsing
+ * table contents.
+ * @param  {ArrayBuffer}
+ * @return {ArrayBuffer}
+ */
+function woff_to_otf(buffer) {
+    if (buffer.constructor !== ArrayBuffer)
+        buffer = new Uint8Array(buffer).buffer;
+    const data = new DataView(buffer, 0)
+      , out = []
+      , signature = parse.getTag(data, 0)
+      ;
+
+    if (signature !== 'wOFF')
+        throw new Error(`TYPE ERROR signature must be wOFF but is: "${signature}"`);
+
+    const flavor = parse.getTag(data, 4)
+      , numTables = parse.getUShort(data, 12)
+      , tableEntries = parseWOFFTableEntries(data, numTables)
+      , max = []
+      ;
+    for (let n = 0; n < 64; n++) {
+        if (Math.pow(2, n) > numTables)
+            break;
+        max.splice(0, Infinity, n, 2 ** n);
+    }
+  const searchRange = max[1] * 16
+    , entrySelector = max[0]
+    , rangeShift = numTables * 16 - searchRange
+    ;
+
+    out.push(
+        ...encode.TAG(flavor)
+      , ...encode.USHORT(numTables)
+      , ...encode.USHORT(searchRange)
+      , ...encode.USHORT(entrySelector)
+      , ...encode.USHORT(rangeShift)
+    );
+    let offset = out.length + numTables * 16;
+
+    for (let i=0; i<numTables; i++) {
+        const tableEntry = tableEntries[i];
+        out.push(
+            ...encode.TAG(tableEntry.tag)
+          , ...encode.ULONG(tableEntry.checksum)
+          , ...encode.ULONG(offset)
+          , ...encode.ULONG(tableEntry.length)
+        )
+        tableEntry.outOffset = offset;
+        offset += tableEntry.length;
+        if ((offset % 4) !== 0)
+            offset += 4 - (offset % 4);
+    }
+    const initialData = new Uint8Array(out.length)
+      , buffers = [initialData]
+      ;
+    for (let i=0,l=out.length; i<l; i++)
+        initialData[i] = out[i];
+
+    for (let i=0; i<numTables; i++) {
+        const tableEntry = tableEntries[i]
+          , table = uncompressTable(data, tableEntry) // => {data: view, offset: 0};
+          , offset = tableEntry.outOffset + tableEntry.length
+          , padding = (offset % 4) !== 0
+                    ? 4 - (offset % 4)
+                    : 0
+            ;
+        buffers.push(
+            new Uint8Array(table.data.buffer, table.offset, tableEntry.length)
+          , new Uint8Array(padding)
+        );
+    }
+    const result = new Uint8Array(buffers.reduce((accum, buffer)=>accum+buffer.byteLength, 0));
+    buffers.reduce((offset, buffer)=>{
+        result.set(buffer, offset)
+        return offset + buffer.byteLength
+    }, 0)
+    return result.buffer;
+}
+
 export {
     Font,
     Glyph,
@@ -543,7 +624,5 @@ export {
     parseBuffer as parse,
     load,
     loadSync,
-    parseWOFFTableEntries,
-    uncompressTable,
     woff_to_otf
 };
